Update 1.3: rename some file and set up sign up router
