+++ src/Components/Button/Button.jsx
@@ -1,4 +1,4 @@
-import { CustomButton } from "./Style-component";
+import { CustomButton } from "./style_component";
 // this button will use for login and sign up
 export default function Buttons({ children }) {
   return <CustomButton type="submit">{children}</CustomButton>;
