+++ src/Components/Header/style_component.jsx
@@ -1,0 +1,13 @@
+import styled from "@emotion/styled";
+
+import background from "../../images/background.png";
+
+export const FormWrapper = styled.div`
+  font-family: sans-serif;
+  display: flex;
+  flex-direction: column;
+  height: 88vh;
+  background-image: url(${background});
+  background-size: cover;
+  background-repeat: no-repeat;
+`;
