+++ src/Components/Search-input/style_component.jsx
