+++ src/index.js
@@ -3,7 +3,8 @@ import ReactDOM from "react-dom/client";
 import { BrowserRouter, Route, Routes, Navigate } from "react-router-dom";
 
 import Index from "./view/index";
-import Login from "./view/log-in/LoginPage";
+import Login from "./view/log-in/login_page";
+import Signup from "./view/sign-up/signup_page";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 
@@ -12,6 +13,7 @@ root.render(
     <Routes>
       <Route path="/" exact element={<Index />} />
       <Route path="/login-page" exact element={<Login />} />
+      <Route path="/sign-up" exact element={<Signup />} />
       <Route path="*" element={<Navigate to="/" replace />} />
     </Routes>
   </BrowserRouter>
