+++ src/view/log-in/login_page.jsx
@@ -1,0 +1,30 @@
+import React, { useState } from "react";
+import LoginForm from "./login_form";
+import NavbarForHomePage from "../../Components/Navbar/navbar_for_home_page";
+
+const LoginPage = () => {
+  const [loggedInUser, setLoggedInUser] = useState(null);
+
+  const handleLogin = (username) => {
+    // This is where you update the login state
+    setLoggedInUser(username);
+  };
+
+  return (
+    <div>
+      {loggedInUser ? (
+        <div>
+          <p>Hello, {loggedInUser}!</p>
+          {/* Add logout button and logic here if needed */}
+        </div>
+      ) : (
+        <>
+          <NavbarForHomePage isSearchVisible={false} />
+          <LoginForm onLogin={handleLogin} />
+        </>
+      )}
+    </div>
+  );
+};
+
+export default LoginPage;
