+++ src/view/log-in/style_component.jsx
@@ -1,0 +1,24 @@
+import styled from "@emotion/styled";
+
+export const CustomLoginForm = styled.div`
+  max-width: 26rem;
+  margin-right: auto;
+  margin-left: auto;
+  margin-top: 145px;
+  padding: 2.25rem 1.25rem 1.25rem 1.25rem;
+  border-radius: 0.5rem;
+  background-color: #f3f7fd;
+  font-family: "Open Sans", sans-serif;
+  text-align: center;
+  width: 100%;
+`;
+
+export const SignUp = styled.a`
+  color: #555;
+  text-decoration: underline;
+`;
+
+export const BirdDecor = styled.img`
+  position: absolute;
+  top: 80px;
+`;
