+++ src/view/sign-up/signup_from.jsx
@@ -1,0 +1,12 @@
+import React, { useState } from "react";
+
+import Header from "../../Components/Header/index";
+
+function SignUpForm() {
+  return (
+    <>
+      <Header />
+    </>
+  );
+}
+export default SignUpForm;
